art_agent_team/agents/upscale_agent.py
import logging
import os
# Potential libraries/APIs for upscaling:
# - ESRGAN implementations (e.g., realesrgan)
# - OpenCV (might have some basic methods)
# - Commercial upscaling APIs (e.g., Upscale.media API)

class UpscaleAgent:
    def __init__(self, config):
        self.config = config
        self.workspace_folder = config.get('workspace_folder', 'workspace') # Use .get for safer config access
        # Placeholder for tool analysis findings and selected libraries/APIs
        # Need to determine optimal libraries/APIs for 4k upscaling that preserve details.

    def upscale_image(self, input_path):
        """
        Upscales the input image to 4k resolution.
        """
        logging.info(f"Attempting to upscale image: {input_path}")
        output_filename = f"upscaled_{os.path.basename(input_path)}"
        output_path = os.path.join(self.workspace_folder, output_filename)

        # Ensure workspace folder exists
        os.makedirs(self.workspace_folder, exist_ok=True)

        # TODO: Implement actual image upscaling using chosen libraries/APIs.
        # This step requires integrating a library or API capable of
        # increasing image resolution to 4k (3840x2160 pixels).
        # The chosen tool should prioritize preserving details like brushwork,
        # texture, and color nuances to maintain the artwork's integrity.

        # Example using a hypothetical upscaling library:
        # try:
        #     upscaled_image = UpscaleLibrary.upscale(input_path, target_resolution=(3840, 2160))
        #     upscaled_image.save(output_path)
        #     logging.info(f"Image upscaled and saved to: {output_path}")
        #     return output_path
        # except Exception as e:
        #     logging.exception(f"Error during image upscaling for {input_path}: {e}")
        #     return None

        # Placeholder: Simulate upscaling by creating a dummy 4k image.
        # REPLACE THIS WITH ACTUAL UPSCALE IMPLEMENTATION.
        try:
            from PIL import Image
            # Simulate a 4k image output
            dummy_upscaled_img = Image.new('RGB', (3840, 2160), color = 'orange')
            dummy_upscaled_img.save(output_path)
            logging.warning(f"Placeholder: Simulated upscaling by creating a dummy 4k image at {output_path} for {input_path}.")
            return output_path
        except Exception as e:
            logging.exception(f"Placeholder error simulating upscaling for {input_path}: {e}")
            return None


if __name__ == "__main__":
    # Example Usage (for testing the agent individually)
    # In the final application, this will be orchestrated by main.py
    sample_config = {
        'workspace_folder': 'workspace' # Relative path
    }
    # Ensure workspace folder exists for standalone testing
    os.makedirs(sample_config['workspace_folder'], exist_ok=True)

    # Use the provided test image path (simulating a cropped image) for testing
    test_cropped_image_path = '/Users/nathaniel.cannon/Documents/VScodeWork/Art_AI/test_image/Tomioka Soichiro - Trees (1961).jpeg'
    # Note: This test image is not 16:9 or low resolution, but serves to test the agent call.
    # A proper test would involve a cropped, lower-res image.

    # Ensure the test image exists
    if not os.path.exists(test_cropped_image_path):
        logging.error(f"Test cropped image not found at: {test_cropped_image_path}")
    else:
        agent = UpscaleAgent(sample_config)
        upscaled_image_path = agent.upscale_image(test_cropped_image_path)
        if upscaled_image_path:
            logging.info(f"Upscale process initiated for: {test_cropped_image_path}")
            logging.info(f"Output path (placeholder simulation): {upscaled_image_path}")
        else:
            logging.error(f"Upscale process failed for: {test_cropped_image_path}")